qsar_qspr_system/requirements.txt
pandas
numpy
scikit-learn
rdkit
biopython
